// client/dist/widgets/common/print/src/config.ts
export const DEFAULT_DPI = 96
export const WKID_LINK = 'https://developers.arcgis.com/rest/services-reference/enterprise/using-spatial-references.htm'
export const CIMMarkerNorthArrow = 'CIMMarkerNorthArrow'

export enum PrintExtentType {
  CurrentMapExtent = 'CURRENT MAP EXTENT',
  CurrentMapScale = 'CURRENT MAP SCALE',
  SetMapScale = 'SET MAP SCALE'
}

export enum MapFrameUnit {
  Point = 'POINT',
  Inch = 'INCH',
  Centimeter = 'CENTIMETER',
  Millimeter = 'MILLIMETER'
}

export interface WebMapFrameSize {
  width: number
  height: number
}

interface CustomTextElementEnable {
  [key: string]: boolean
}

export const DEFAULT_TEMPLATE_INFO = {
  layoutOptions: {
    customTextElements: [],
    hasAuthorText: true,
    hasCopyrightText: true,
    hasLegend: true,
    hasTitleText: true
  },
  mapSurroundInfos: [],
  pageSize: null,
  pageUnits: null,
  webMapFrameSize: null
}

export const DEFAULT_COMMON_SETTING = {
  scalePreserved: false,
  outScale: 36978595.474472,
  layoutOptions: {
    titleText: 'ArcGIS Web Map'
  },
  exportOptions: {
    dpi: DEFAULT_DPI
  },
  printExtentType: PrintExtentType.CurrentMapExtent,
  attributionVisible: false,
  forceFeatureAttributes: true,
  wkid: 102100,
  wkidLabel: 'WGS_1984_Web_Mercator_Auxiliary_Sphere',
  enableTitle: true,
  legendEnabled: true,
  enableMapPrintExtents: true,
  enableOutputSpatialReference: true,
  enableQuality: true,
  enableFeatureAttribution: true,
  enableMapSize: true,
  overrideCommonSetting: false,
  enableAuthor: true,
  enableCopyright: true,
  enableLegend: true,
  enableScalebarUnit: true,
  enableCustomTextElements: true
}

export enum ModeType {
  Classic = 'CLASSIC',
  Compact = 'COMPACT'
}

export enum PrintServiceType {
  OrganizationService = 'ORGANIZATION SERVICE',
  Customize = 'CUSTOMIZE'
}

export enum PrintTemplateType {
  OrganizationTemplate = 'ORGANIZATION TEMPLATE',
  Customize = 'CUSTOMIZE'
}
// ...
export type FormatType = 'pdf' | 'png32' | 'png8' | 'jpg' | 'gif' | 'eps' | 'svg' | 'svgz'
// ...
export interface LayoutInfo {
  layoutTemplate: LayoutType
  pageSize?: number[]
  mapFrameSize?: number[]
  mapFrameUnit?: MapFrameUnit
  hasAuthorText?: boolean
  hasCopyrightText?: boolean
  hasLegend?: boolean
  hasTitleText?: boolean
  layout: LayoutType
  layoutOptions: {
    hasAuthorText?: boolean
    hasCopyrightText?: boolean
    hasLegend?: boolean
    hasTitleText?: boolean
    customTextElements?: any[]
  }
}

export type PrintParameters = __esri.PrintParameters
export type MapView = __esri.MapView

export interface PrintTemplateProperties extends __esri.PrintTemplateProperties {
  templateId?: string
  printExtentType?: PrintExtentType
  wkid?: number | string
  label?: string
  legendEnabled?: boolean

  //Size of map in template
  mapFrameSize?: number[]
  mapFrameUnit?: MapFrameUnit

  overrideCommonSetting?: boolean
  enableTitle?: boolean
  enableMapPrintExtents?: boolean
  enableOutputSpatialReference?: boolean
  enableQuality?: boolean
  enableMapSize?: boolean
  enableFeatureAttribution?: boolean
  enableAuthor?: boolean
  enableCopyright?: boolean
  enableLegend?: boolean
  enableScalebarUnit?: boolean
  enableMapAttribution?: boolean
  enableCustomTextElements?: boolean
  customTextElementEnableList?: CustomTextElementEnable[]
  enableNorthArrow?: boolean

  hasAuthorText?: boolean
  hasCopyrightText?: boolean
  hasLegend?: boolean
  hasTitleText?: boolean
  selectedFormatList?: FormatType[]

  wkidLabel?: string
}

export type IMPrintTemplateProperties = ImmutableObject<PrintTemplateProperties>
export interface config {
  modeType: ModeType
  printServiceType: PrintServiceType
  printTemplateType: PrintTemplateType
  useUtility?: UseUtility
  printCustomTemplate?: PrintTemplateProperties[]
  printOrgTemplate?: PrintTemplateProperties[]
  commonSetting?: PrintTemplateProperties
  formatList?: FormatType[]
  defaultFormat?: FormatType
  layoutChoiceList?: LayoutInfo[]
  hasInitBorder?: boolean
}

export type IMConfig = ImmutableObject<config>